Moved dev rpc server to rpcservers folder
The build tag name got changed?
