+++ rpcservers/dev/driver.go
@@ -1,7 +1,7 @@
 //go:build dev
 // +build dev
 
-package devrpc
+package dev
 
 import (
 	"fmt"
