+++ subrpcserver_config.go
@@ -13,7 +13,6 @@ import (
 	"github.com/lightningnetwork/lnd/htlcswitch"
 	"github.com/lightningnetwork/lnd/invoices"
 	"github.com/lightningnetwork/lnd/lncfg"
-	"github.com/lightningnetwork/lnd/lnrpc/devrpc"
 	"github.com/lightningnetwork/lnd/lnrpc/invoicesrpc"
 	"github.com/lightningnetwork/lnd/lnrpc/neutrinorpc"
 	"github.com/lightningnetwork/lnd/lnrpc/peersrpc"
@@ -28,6 +27,7 @@ import (
 	"github.com/lightningnetwork/lnd/routing"
 	autopilotrpc "github.com/lightningnetwork/lnd/rpcservers/autopilot"
 	chainrpc "github.com/lightningnetwork/lnd/rpcservers/chain"
+	devrpc "github.com/lightningnetwork/lnd/rpcservers/dev"
 	"github.com/lightningnetwork/lnd/sweep"
 	"github.com/lightningnetwork/lnd/watchtower"
 	"github.com/lightningnetwork/lnd/watchtower/wtclient"
